requirements.txt
fastapi==0.104.1
httpx==0.25.2
openai==1.3.0
pydantic==2.5.0
pytest-asyncio==0.21.1
pytest==7.4.3
python-dotenv==1.0.0
redis==5.0.1
uvicorn[standard]==0.24.0